% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_const.R
\name{:=}
\alias{:=}
\title{Create Constant}
\usage{
x := value
}
\arguments{
\item{x}{a constant name}

\item{value}{a value to be assigned to x.}
}
\description{
Assign a value to a constant. Once a constant has been
created, it cannot be re-assigned.
}
